Increase program size to 2 bytes

+++ src/main.c
@@ -6,7 +6,7 @@
 #define ADD_INSTRUCTION(...) add_instruction((uint8_t[]){__VA_ARGS__}, sizeof((uint8_t[]){__VA_ARGS__}))
 #define ADD_PROGRAM(addr,...) add_program(addr, (uint8_t[]){__VA_ARGS__}, sizeof((uint8_t[]){__VA_ARGS__}))
 
-void add_instruction(uint8_t *instructions, uint8_t size)
+void add_instruction(uint8_t *instructions, uint16_t size)
 {
     static uint16_t ram_pos = 0x0600;
     memcpy(mainMemory + ram_pos, instructions, size);
@@ -14,7 +14,7 @@ void add_instruction(uint8_t *instructions, uint8_t size)
     printf("RAM SIZE: %x\n", ram_pos);
 }
 
-void add_program(uint16_t addr, uint8_t *program, uint8_t size)
+void add_program(uint16_t addr, uint8_t *program, uint16_t size)
 {
     memcpy(mainMemory + addr, program, size);
 }
